+++ CHANGELOG.md
@@ -28,6 +28,13 @@ The general format is:
 - C from D
 
 ```
+# 2.2.0 - DATE (17/02/2022)
+
+
+### Added
+
+- Add support for sharing the IP of one Task to another Task, via env variables
+- support for setting args instead of command
 # 2.1.1 - DATE (16/02/2022)